Initially added shuggest script and styles
This script is able to decorate all existing shell command textareas
with an infobox. On click the path will be added to the nearby textarea
with $JENKINS_HOME/userContent/ path prefix.

Next to the path a description is displayed and upon hovering the [?]
icon an usage info tooltip is shown. When clicking [show] the file
contents will be shown in a new tab.